Implement continue statement
We pass an extra argument to the while statement (increment expr)
this allow us to know if we are executing a "for loop".

When a continue statement is reached we execute the increment Expr
this prevents an infinite loop.

There is probably a better way but for now this seems to work.